Close to having a general method for segmenting by paragraph
I added a SegmentedDocument class, which encodes higher-level document
structure in a way that should be easy to index.  I also refactored
BuildCorpusIndex a bit, so that the three different ways of specifying a
corpus all use the same code path, removing a fair amount of code duplication.
This should make it easier to add an option for using the SegmentedDocument
class that I wrote.

I also added some tests for the new code that I wrote.